requirements.txt
pandas==2.0.2
argparse==1.4.0
requests==2.31.0
PyMySQL==1.0.3
setuptools==67.8.0
mysql-connector-python==8.0.33
pyhumps==3.8.0